chore: rename resources folder
