+++ packages/experimental/src/components/AboutModal/AboutModal-test.js
@@ -10,10 +10,10 @@ import React from 'react';
 import { Link } from 'carbon-components-react';
 
 import { AboutModal } from './AboutModal';
-import ExampleLogo from './_resources/example-logo.svg';
-import ansibleLogo from './_resources/ansible-logo.png';
-import grafanaLogo from './_resources/grafana-logo.png';
-import jsLogo from './_resources/js-logo.png';
+import ExampleLogo from './_story-assets/example-logo.svg';
+import ansibleLogo from './_story-assets/ansible-logo.png';
+import grafanaLogo from './_story-assets/grafana-logo.png';
+import jsLogo from './_story-assets/js-logo.png';
 
 const content = 'This is example content';
 const logo = (
