+++ packages/experimental/src/components/AboutModal/_story-assets/ansible-logo.png
