+++ packages/experimental/src/components/AboutModal/_story-assets/grafana-logo.png
