+++ packages/experimental/src/components/AboutModal/_story-assets/js-logo.png
